# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  waiter: "waiter",
  kitchen: "kitchen",
  accountant: "accountant",
  owner: "owner",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  waiter: "waiter",
  kitchen: "kitchen",
  accountant: "accountant",
  owner: "owner",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  waiter: "waiter",
  kitchen: "kitchen",
  accountant: "accountant",
  owner: "owner",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  role?: UpdateUserBodyRole;
  password?: string;
}

export type TableStatus = (typeof TableStatus)[keyof typeof TableStatus];

export const TableStatus = {
  available: "available",
  occupied: "occupied",
  cooking: "cooking",
  ready: "ready",
  billing: "billing",
} as const;

export interface Table {
  id: number;
  number: number;
  capacity: number;
  status: TableStatus;
  /** @nullable */
  currentOrderId?: number | null;
}

export interface CreateTableBody {
  number: number;
  capacity: number;
}

export type UpdateTableBodyStatus =
  (typeof UpdateTableBodyStatus)[keyof typeof UpdateTableBodyStatus];

export const UpdateTableBodyStatus = {
  available: "available",
  occupied: "occupied",
  cooking: "cooking",
  ready: "ready",
  billing: "billing",
} as const;

export interface UpdateTableBody {
  number?: number;
  capacity?: number;
  status?: UpdateTableBodyStatus;
}

export interface TablesSummary {
  available: number;
  occupied: number;
  cooking: number;
  ready: number;
  billing: number;
  total: number;
}

export interface MenuItem {
  id: number;
  name: string;
  category: string;
  price: number;
  /** @nullable */
  description?: string | null;
  available: boolean;
  createdAt: string;
}

export interface CreateMenuItemBody {
  name: string;
  category: string;
  price: number;
  description?: string;
  available?: boolean;
}

export interface UpdateMenuItemBody {
  name?: string;
  category?: string;
  price?: number;
  description?: string;
  available?: boolean;
}

export interface TopSellingItem {
  menuItemId: number;
  name: string;
  category: string;
  totalQuantity: number;
  totalRevenue: number;
}

export interface OrderItem {
  id: number;
  orderId: number;
  menuItemId: number;
  menuItemName: string;
  menuItemPrice: number;
  quantity: number;
  /** @nullable */
  notes?: string | null;
  subtotal: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  billing: "billing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  tableId: number;
  tableNumber: number;
  waiterId: number;
  waiterName: string;
  status: OrderStatus;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderItemInput {
  menuItemId: number;
  quantity: number;
  notes?: string;
}

export interface CreateOrderBody {
  tableId: number;
  items: CreateOrderItemInput[];
  notes?: string;
}

export type UpdateOrderBodyStatus =
  (typeof UpdateOrderBodyStatus)[keyof typeof UpdateOrderBodyStatus];

export const UpdateOrderBodyStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  billing: "billing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderBody {
  status?: UpdateOrderBodyStatus;
  items?: CreateOrderItemInput[];
  notes?: string;
}

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

export const PaymentMethod = {
  cash: "cash",
  upi: "upi",
  card: "card",
} as const;

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  completed: "completed",
} as const;

export interface Payment {
  id: number;
  orderId: number;
  tableNumber: number;
  subtotal: number;
  tax: number;
  discount: number;
  totalAmount: number;
  method: PaymentMethod;
  status: PaymentStatus;
  createdAt: string;
}

export type CreatePaymentBodyMethod =
  (typeof CreatePaymentBodyMethod)[keyof typeof CreatePaymentBodyMethod];

export const CreatePaymentBodyMethod = {
  cash: "cash",
  upi: "upi",
  card: "card",
} as const;

export interface CreatePaymentBody {
  orderId: number;
  discount?: number;
  method: CreatePaymentBodyMethod;
  taxRate?: number;
}

export interface DailyReport {
  date: string;
  totalRevenue: number;
  totalOrders: number;
  totalCustomers: number;
  cashRevenue: number;
  upiRevenue: number;
  cardRevenue: number;
  avgOrderValue: number;
}

export interface DailyRevenue {
  date: string;
  revenue: number;
  orders: number;
}

export interface InventoryItem {
  id: number;
  name: string;
  quantity: number;
  unit: string;
  lowStockThreshold: number;
  isLowStock: boolean;
  updatedAt: string;
}

export interface CreateInventoryItemBody {
  name: string;
  quantity: number;
  unit: string;
  lowStockThreshold: number;
}

export interface UpdateInventoryItemBody {
  name?: string;
  quantity?: number;
  unit?: string;
  lowStockThreshold?: number;
}

export interface Expense {
  id: number;
  description: string;
  amount: number;
  category: string;
  date: string;
  createdAt: string;
}

export interface CreateExpenseBody {
  description: string;
  amount: number;
  category: string;
  date: string;
}

export interface UpdateExpenseBody {
  description?: string;
  amount?: number;
  category?: string;
  date?: string;
}

export interface OwnerDashboard {
  todayRevenue: number;
  todayOrders: number;
  todayCustomers: number;
  monthRevenue: number;
  monthOrders: number;
  tableOccupancyRate: number;
  totalExpensesThisMonth: number;
  netProfitThisMonth: number;
  lowStockCount: number;
}

export interface StaffPerformance {
  userId: number;
  name: string;
  role: string;
  totalOrders: number;
  totalRevenue: number;
}

export type ListMenuItemsParams = {
  category?: string;
  available?: boolean;
  search?: string;
};

export type GetTopSellingItemsParams = {
  limit?: number;
};

export type ListOrdersParams = {
  status?: string;
  tableId?: number;
  date?: string;
};

export type ListPaymentsParams = {
  date?: string;
  method?: string;
};

export type GetDailyReportParams = {
  date?: string;
};

export type GetMonthlyChartParams = {
  year?: number;
  month?: number;
};

export type ListExpensesParams = {
  month?: number;
  year?: number;
};