# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Tables
...
  /menu-items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu-items/categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List unique menu categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /menu-items/top-selling:
    get:
      operationId: getTopSellingItems
      tags: [menu]
      summary: Get top selling menu items
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top selling items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopSellingItem"
  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add an inventory item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/low-stock:
    get:
      operationId: listLowStockItems
      tags: [inventory]
      summary: Get low stock alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Low stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Expenses
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Add an expense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Users / Staff
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # User
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [waiter, kitchen, accountant, owner]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        description:
          type: ["string", "null"]
        available:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - price
        - available
        - createdAt
    CreateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        description:
          type: string
        available:
          type: boolean
      required:
        - name
        - category
        - price
    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        description:
          type: string
        available:
          type: boolean
    TopSellingItem:
      type: object
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        category:
          type: string
        totalQuantity:
          type: integer
        totalRevenue:
          type: number
      required:
        - menuItemId
        - name
        - category
        - totalQuantity
        - totalRevenue

    # Order
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        lowStockThreshold:
          type: number
        isLowStock:
          type: boolean
        updatedAt:
          type: string
      required:
        - id
        - name
        - quantity
        - unit
        - lowStockThreshold
        - isLowStock
        - updatedAt
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        lowStockThreshold:
          type: number
      required:
        - name
        - quantity
        - unit
        - lowStockThreshold
    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        lowStockThreshold:
          type: number

    # Expenses
    Expense:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - description
        - amount
        - category
        - date
        - createdAt
    CreateExpenseBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
      required:
        - description
        - amount
        - category
        - date
    UpdateExpenseBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string

    # Analytics
...